fix: Moved from unit tests to integration tests, now with proper dependency checks
- Introduced integration tests for the `FfmpegAudioPreprocessor` to validate its behavior when FFmpeg is available on the system.
- Tests cover scenarios such as checking availability, preprocessing audio files with options to replace originals or create new files, and returning processing statistics.
- Moved relevant tests from unit tests to integration tests to ensure proper execution context and compliance with architectural standards.
- Enhanced test coverage for audio processing features, ensuring robust validation of the audio preprocessing capabilities.
